Revamp permissions system
This commit changes how permissions work. It adds a `requireRoles` field
to commands. If a command has this defined, the bot will check that the
user has a role in the list before allowing them to run the command.

This commit also adds a `REQUIRED_TO_USE_BOT` and a `BANNED_FROM_BOT`
role list. If the former has any roles in it, the bot will check that a
user running a command has one of the roles before running the command,
ignoring the message if not. The latter contains roles that are to be
ignored by the bot. If a user has a role in the banned list, their ban
supersedes the `REQUIRED_TO_USE_BOT` check.