Delete message in #welcome-room when a user leaves
This commit adds code to the leave handler that looks at the past 100
messages in #welcome-room and deletes messages that the bot sent that
mention the leaving user. This is intended to delete the welcome message
but as a side effect will also delete any other message that mentions
the user (eg a `!slap` response). I don't think this is a big deal.