Fix not removing other regions with !setregion
This is because calling removeRoles before addRole seems to have no
effect. Even when addRole is placed in the .then() for removeRoles the
role removal has no effect. I will be investigating this further but
until then I've reverted back to the behavior of re-building the user's
role list with the changes in place.