Add `onlyIn` option to commands
Setting this on a command will only allow it to be used in a channel by
that name. If `allowDM` is set, it can also be used in a DM.

If this option is set and a user uses the command outside of its
restricted channel, the bot will delete the message and DM the user a
list of channels that the command may be used in. If the user does not
have 'READ_MESSAGES' permission in the channel, the user will not be
told about the channel. If no channels are valid for a user or server,
the bot will just say that the "command isn't valid in that channel".