Fix !joined command using bad member to get date
Previously, we were using the `GuildMember.joinDate` member to get the
date that a user joined. This doesn't exist in the discord.js 9 docs but
must have been a deprecated or undocumented member? This is gone now in
discord.js 10 so we were just passing `undefined` to `moment()`. This
just changes the command to use the `GuildMember.joinDate` member.