Remove 'Member' role when a user is timed out
Because 'ALLOW' trumps 'DENY' in Discord, we need to remove the user's
membership in order for 'Restricted' to take effect.

This is only a partial fix, we still need to be able to handle giving
the user back their membership after their restriction is over.

This is a partial fix for gaymers-discord/DiscoBot#152