Handles error in deleted message event (#216)
* Use a defined AFK channel on restrict

Previously, we would search for a channel named 'AFK' to move a
restricted user to. This instead first looks at the server's defined AFK
channel and if that can't be found, then reverts to looking for a
channel named 'AFK'. Failing that, it will mute the user.

* Handles error in deleted message event.

* Fix null defref for guild member on deletion

This also adds a fallback for usernames of users that left