Add a !quote command
This command quotes a message when given a message ID.

Note that this only uses discord.js's local message cache. If the bot
does not have the desired message in its cache, it appears to not exist.
There may be a way to ask discord.js to hit the Discord API for message
history, I'll look into this in the future.

Closes gaymers-discord/DiscoBot#226

:heart: